Replace Int by Double on the interfacing methods of RoomVisual.

+++ src/main/kotlin/screeps/api/RoomVisual.kt
@@ -3,20 +3,20 @@ package screeps.api
 external class RoomVisual(roomName: String) {
     val roomName: String
 
-    fun line(x1: Int, y1: Int, x2: Int, y2: Int, style: LineStyle = definedExternally): RoomVisual
+    fun line(x1: Double, y1: Double, x2: Double, y2: Double, style: LineStyle = definedExternally): RoomVisual
     fun line(start: RoomPosition, end: RoomPosition, style: LineStyle = definedExternally): RoomVisual
-    fun line(x: Int, y: Int, style: LineStyle = definedExternally): RoomVisual
+    fun line(x: Double, y: Double, style: LineStyle = definedExternally): RoomVisual
     fun line(pos: RoomPosition, style: LineStyle = definedExternally): RoomVisual
-    fun circle(x: Int, y: Int, style: CircleStyle = definedExternally): RoomVisual
+    fun circle(x: Double, y: Double, style: CircleStyle = definedExternally): RoomVisual
     fun circle(pos: RoomPosition, style: CircleStyle = definedExternally): RoomVisual
-    fun rect(x: Int, y: Int, width: Int, height: Int, style: ShapeStyle = definedExternally): RoomVisual
-    fun rect(topLeftPosition: RoomPosition, width: Int, height: Int, style: ShapeStyle = definedExternally): RoomVisual
+    fun rect(x: Double, y: Double, width: Double, height: Double, style: ShapeStyle = definedExternally): RoomVisual
+    fun rect(topLeftPosition: RoomPosition, width: Double, height: Double, style: ShapeStyle = definedExternally): RoomVisual
     fun poly(points: Array<RoomPosition>, style: ShapeStyle = definedExternally): RoomVisual
-    fun poly(points: Array<Array<Int>>, style: ShapeStyle = definedExternally): RoomVisual
-    fun text(text: String, x: Int, y: Int, style: TextStyle = definedExternally): RoomVisual
+    fun poly(points: Array<Array<Double>>, style: ShapeStyle = definedExternally): RoomVisual
+    fun text(text: String, x: Double, y: Double, style: TextStyle = definedExternally): RoomVisual
     fun text(text: String, position: RoomPosition, style: TextStyle = definedExternally): RoomVisual
     fun clear(): RoomVisual
-    fun getSize(): Int
+    fun getSize(): Double
 
     interface Style : Options {
         var opacity: Double?
